allow using global python2 and python3 executable

+++ autoload/python2_support.sh
@@ -7,11 +7,19 @@ cd $(dirname "${BASH_SOURCE[0]}")
 
 echo 'initializing python2 virtualenv for neovim'
 
-rm -rf nvim_py2
-python2 -m virtualenv nvim_py2
-. nvim_py2/bin/activate
-pip install "$@"
-deactivate
+venv=$1
+
+shift
+
+if [[ $venv == 1 ]]
+then
+    rm -rf nvim_py2
+    python2 -m virtualenv nvim_py2
+    . nvim_py2/bin/activate
+    pip install "$@"
+else
+    python2 -m pip install --user "$@"
+fi
 
 echo 'init finished'
